+++ components/FormWithSubmit.js
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import {
   StyleSheet,
   Text,
@@ -13,18 +13,32 @@ import Colors from "../constants/Colors";
 import { TextInput } from "react-native";
 import { AsyncStorage } from "react-native";
 
+const statusDict = {
+    applied: "Applied",
+    phoneInterview: "Phone interview",
+    second_interview: "Second Interview",
+    final_interview: "Final Interview",
+    take_home_test: "Take home test",
+    offer: "Offer",
+    rejected: "Rejected",
+    cancelled: "Cancelled",
+  };
+
+
 const statusOptions = Object.entries(statusDict).map(([key, value]) => (
     <Picker.Item label={value} value={key} key={key} />
   ));
   
-  export const FormWithSubmit = ( formValues ) => {
+  export function FormWithSubmit( formValues ) {
     const [updatedFormValues, setUpdatedFormValues] = useState({});
     const [jobID, setJobID] = useState("");
     const [interestLevel, setInterestLevel] = useState(0);
   
-    if (formValues != null ) {
-        setUpdatedFormValues(formValues)
-    }
+    useEffect(() => {
+        if (formValues != null) {
+                setUpdatedFormValues(formValues)
+            }
+      }, [formValues]);
 
     function onSubmit() {
       if (jobID != undefined && jobID != "") {
@@ -121,3 +135,44 @@ const statusOptions = Object.entries(statusDict).map(([key, value]) => (
       </ScrollView>
     );
   };
+
+  const styles = StyleSheet.create({
+    container: {
+      flex: 1,
+      backgroundColor: "white",
+      marginRight: 10,
+      marginLeft: 10,
+    },
+    contentContainer: {
+      paddingTop: 15,
+    },
+    textInput: {
+      flex: 1,
+      borderBottomWidth: 1,
+      borderRightWidth: 1,
+      borderTopWidth: 1,
+      borderLeftWidth: 1,
+      marginRight: 10,
+      marginLeft: 10,
+      height: 40,
+      padding: 5,
+    },
+    text: {
+      flex: 1,
+      marginRight: 10,
+      marginLeft: 10,
+      height: 40,
+      padding: 5,
+    },
+    button: {
+      flex: 1,
+      alignItems: "center",
+      backgroundColor: Colors.orange,
+      padding: 15,
+      marginRight: 100,
+      marginLeft: 100,
+    },
+    buttonText: {
+      color: "white",
+    },
+  });
